requirements.txt
torch
git+https://github.com/huggingface/diffusers
huggingface-hub
Pillow
flask
controlnet-aux
numpy
opencv-python
runpod
transformers
requests